- `JsonHelper`クラスを新規追加し、JSON文字列の配列デシリアライズを共通化 - `AddressableMasterDataLoader`クラスを新規追加し、Addressablesのロード処理を分離 - `MasterDataTextAssetLoaderBase`クラス内のロードロジックをリファクタリングし、`JsonHelper`と`AddressableMasterDataLoader`を利用するよう変更 - 不必要なコードとフィールドを削除し、コードの簡素化と保守性を向上
